requirements.txt
python-mpd2==3.1.1
smart_open[http]==7.1.0
paho-mqtt==2.1.0
inky==2.0.0
pillow==11.1.0
flask==3.1.0
colorthief==0.2.1
structlog==25.1.0
numpy==2.2.3
